A-Video: Automate capture test
This patch automates the sensor capture test. Before this patch
the test required user input to determine pass/fail. After this
patch pass/fail is determined automatically based on golden files.
The patch requires adding a 'sensor_box' => "<box id>" entry to
the dut params field in the bench file so that appropiate golden
files are used during the test

+++ lib/utils.rb
@@ -137,11 +137,12 @@ def use_memory(mem, sys=@equipment['dut1'])
                  1
                end
    available_mem = mem_available.to_i * multiplier - mem
-   return if available_mem < 1
-   sys.send_cmd("memtester #{available_mem}B &>/dev/null &",sys.prompt)
-   sleep 5 # wait for system to allocate memory
+   if available_mem > 0
+     sys.send_cmd("memtester #{available_mem}B &>/dev/null &",sys.prompt)
+     sleep 5 # wait for system to allocate memory
+   end
    yield
-   sys.send_cmd("killall -9 memtester")
+   sys.send_cmd("killall -9 memtester") if available_mem > 0
    sleep 5 # wait for system to reclaim memory
    rescue Exception => e
      puts e.to_s
